FIrst Unit tests

+++ src/main/java/Main.java
@@ -5,7 +5,7 @@ import com.utils.Utils;
 public class Main {
     public static void main(String[] args) {
 
-        int[] table ={1,2,3,4,5};
+        double[] table ={1,2,3,4,5};
         System.out.println(Utils.add(table));
 
         //dirty casting to avoid and try to test it
