+++ src/main/java/com/utils/Utils.java
@@ -3,9 +3,9 @@ package com.utils;
 // Static methods to use this class
 public class Utils {
 
-    public static double add(int[] aValues){
+    public static double add(double[] aValues){
         double result=0.0;
-        for(int values : aValues){
+        for(double values : aValues){
             result+=values;
         }
         return result;
